backend/requirements.txt
fastapi==0.115.6
mangum==0.19.0
pydantic==2.10.4
pydantic-settings==2.7.1
boto3==1.36.4
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
pdfplumber==0.11.4
python-multipart==0.0.20
httpx==0.28.1
uvicorn==0.34.0
